fixed following and moved some functions into ai script

+++ Behaviors/VikingAI/Fleeing.cs
@@ -6,13 +6,13 @@ public partial class VikingAI
 {
     public bool UpdateFlee(float dt)
     {
-        if (m_fleeIfNotAlerted && !HuntPlayer() && m_targetCreature && !IsAlerted() && Vector3.Distance(m_targetCreature.transform.position, transform.position) - (double)m_targetCreature.GetRadius() > m_alertRange)
+        if (m_fleeIfNotAlerted && !HuntPlayer() && m_targetCreature && !IsAlerted() && Vector3.Distance(m_targetCreature.transform.position, transform.position) - m_targetCreature.GetRadius() > m_alertRange)
         {
             Flee(dt, m_targetCreature.transform.position);
             return true;
         }
         
-        if (m_fleeIfLowHealth > 0.0 && m_timeSinceHurt < (double)m_fleeTimeSinceHurt && m_targetCreature != null &&
+        if (m_fleeIfLowHealth > 0.0 && m_timeSinceHurt < m_fleeTimeSinceHurt && m_targetCreature != null &&
             m_character.GetHealthPercentage() < (double)m_fleeIfLowHealth)
         {
             Flee(dt, m_targetCreature.transform.position);
